requirements.txt
aiohttp>=3.8.0
tortoise-orm>=0.19.0
uvloop>=0.17.0; sys_platform != 'win32'
pyjwt>=2.4.0
zstandard>=0.18.0
lz4>=4.0.0
python-dotenv>=0.19.0
cryptography~=44.0.0
bcrypt~=4.2.1
aioconsole~=0.8.1